Track hovered item and clear hover ui as game state changes
This ensures that note tooltips are not orphaned, e.g. by hovering
over a card, then advancing the replay to when it is played, and
then rewinding to before it was played.

When viewing an in-game replay, incoming server messages about the
current state of the game do not trigger this behavior.